+++ tools/cli/internal/changelog/changelog.go
@@ -15,19 +15,16 @@
 package changelog
 
 import (
-	"encoding/json"
-
+	"github.com/mongodb/openapi/tools/cli/internal/changelog/outputfilter"
 	"github.com/mongodb/openapi/tools/cli/internal/openapi"
 	"github.com/tufin/oasdiff/checker"
 	"github.com/tufin/oasdiff/diff"
-	"github.com/tufin/oasdiff/formatters"
 	"github.com/tufin/oasdiff/load"
 )
 
 const (
-	deprecationDaysStable = 365  //  min days required between deprecating a stable resource and removing it
-	deprecationDaysBeta   = 365  //  min days required between deprecating a beta resource and removing it
-	lan                   = "en" // language for localized output
+	deprecationDaysStable = 365 //  min days required between deprecating a stable resource and removing it
+	deprecationDaysBeta   = 365 //  min days required between deprecating a beta resource and removing it
 )
 
 var breakingChangesAdditionalCheckers = []string{
@@ -38,17 +35,6 @@ var breakingChangesAdditionalCheckers = []string{
 	"response-mediatype-enum-value-removed",
 	"request-body-enum-value-removed",
 	"api-schema-removed",
-}
-
-type Entry struct {
-	ID          string `json:"id"`
-	Text        string `json:"text"`
-	Level       int    `json:"level"`
-	Operation   string `json:"operation"`
-	OperationID string `json:"operationId"`
-	Path        string `json:"path"`
-	Source      string `json:"source"`
-	Section     string `json:"section"`
 }
 
 type Metadata struct {
@@ -84,7 +70,7 @@ func NewMetadata(base, revision, exceptionFilePath string) (*Metadata, error) {
 	}, nil
 }
 
-func (s *Metadata) Check() ([]*Entry, error) {
+func (s *Metadata) Check() ([]*outputfilter.Entry, error) {
 	diffResult, err := s.OasDiff.NewDiffResult()
 	if err != nil {
 		return nil, err
@@ -94,30 +80,7 @@ func (s *Metadata) Check() ([]*Entry, error) {
 		s.Config,
 		diffResult.Report,
 		diffResult.SourceMap,
-
 		checker.INFO)
 
-	return newChangelogEntries(changes, diffResult.SpecInfoPair)
-}
-
-func newChangelogEntries(checkers checker.Changes, specInfoPair *load.SpecInfoPair) ([]*Entry, error) {
-	formatter, err := formatters.Lookup("json", formatters.FormatterOpts{
-		Language: lan,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	bytes, err := formatter.RenderChangelog(checkers, formatters.RenderOpts{}, specInfoPair)
-	if err != nil {
-		return nil, err
-	}
-
-	var entries []*Entry
-	err = json.Unmarshal(bytes, &entries)
-	if err != nil {
-		return nil, err
-	}
-
-	return transformEntries(entries), nil
+	return outputfilter.NewChangelogEntries(changes, diffResult.SpecInfoPair)
 }
