+++ tools/cli/internal/changelog/outputfilter/message_test.go
@@ -1,4 +1,4 @@
-package changelog
+package outputfilter
 
 import (
 	"reflect"
@@ -45,7 +45,7 @@ func TestTransformMessagesInTextField(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			actual := transformMessagesInTextField(tt.entry)
+			actual := transformMessage(tt.entry)
 			if !reflect.DeepEqual(actual, tt.expected) {
 				t.Errorf("transformMessagesInTextField() = %v, want %v", actual, tt.expected)
 			}
